+++ libs/domain/cart/totals/components/component.def.ts
@@ -1,0 +1,6 @@
+export * from './delivery/delivery.def';
+export * from './discount/discount.def';
+export * from './expense/expense.def';
+export * from './subtotal/subtotal.def';
+export * from './tax/tax.def';
+export * from './total/total.def';
