+++ libs/domain/cart/totals/components/delivery/delivery.schema.ts
@@ -1,0 +1,6 @@
+import { ContentComponentSchema } from '@spryker-oryx/experience';
+
+export const cartTotalsDeliveryComponentSchema: ContentComponentSchema = {
+  name: 'Cart totals delivery',
+  group: 'Cart',
+};
