+++ libs/domain/cart/totals/components/delivery/index.ts
@@ -1,0 +1,2 @@
+export * from './delivery.component';
+export * from './delivery.schema';
