+++ libs/domain/cart/totals/components/discount/discount.def.ts
@@ -1,0 +1,21 @@
+import { componentDef } from '@spryker-oryx/core';
+
+const rules = () =>
+  import('../cart-totals.styles').then((m) => m.cartTotalsStyles);
+
+export const cartTotalsDiscountComponent = componentDef({
+  name: 'oryx-cart-totals-discount',
+  impl: () =>
+    import('./discount.component').then((m) => m.CartTotalsDiscountComponent),
+  schema: () =>
+    import('./discount.schema').then(
+      (m) => m.cartTotalsDiscountComponentSchema
+    ),
+  stylesheets: [
+    { rules },
+    {
+      rules: () =>
+        import('./discount.styles').then((m) => m.cartTotalsDiscountStyles),
+    },
+  ],
+});
