+++ libs/domain/cart/totals/components/discount/index.ts
@@ -1,0 +1,2 @@
+export * from './discount.component';
+export * from './discount.schema';
