+++ libs/domain/cart/totals/components/expense/expense.def.ts
@@ -1,0 +1,13 @@
+import { componentDef } from '@spryker-oryx/core';
+
+const rules = () =>
+  import('../cart-totals.styles').then((m) => m.cartTotalsStyles);
+
+export const cartTotalsExpenseComponent = componentDef({
+  name: 'oryx-cart-totals-expense',
+  impl: () =>
+    import('./expense.component').then((m) => m.CartTotalsExpenseComponent),
+  schema: () =>
+    import('./expense.schema').then((m) => m.cartTotalsExpenseComponentSchema),
+  stylesheets: [{ rules }],
+});
