+++ libs/domain/cart/totals/components/expense/expense.schema.ts
@@ -1,0 +1,6 @@
+import { ContentComponentSchema } from '@spryker-oryx/experience';
+
+export const cartTotalsExpenseComponentSchema: ContentComponentSchema = {
+  name: 'Cart totals expense',
+  group: 'Cart',
+};
