+++ libs/domain/cart/totals/components/expense/index.ts
@@ -1,0 +1,2 @@
+export * from './expense.component';
+export * from './expense.schema';
