+++ libs/domain/cart/totals/components/index.ts
@@ -1,4 +1,5 @@
-export * from './cart-total.component';
-export * from './cart-total.model';
-export * from './cart-total.schema';
-export * from './cart-total.styles';
+export * from './discount';
+export * from './expense';
+export * from './subtotal';
+export * from './tax';
+export * from './total';
