+++ libs/domain/cart/totals/components/subtotal/index.ts
@@ -1,0 +1,2 @@
+export * from './subtotal.component';
+export * from './subtotal.schema';
