+++ libs/domain/cart/totals/components/subtotal/subtotal.def.ts
@@ -1,0 +1,17 @@
+import { componentDef } from '@spryker-oryx/core';
+
+export const cartTotalsSubtotalComponent = componentDef({
+  name: 'oryx-cart-totals-subtotal',
+  impl: () =>
+    import('./subtotal.component').then((m) => m.CartTotalsSubtotalComponent),
+  schema: () =>
+    import('./subtotal.schema').then(
+      (m) => m.cartTotalsSubtotalComponentSchema
+    ),
+  stylesheets: [
+    {
+      rules: () =>
+        import('../cart-totals.styles').then((m) => m.cartTotalsStyles),
+    },
+  ],
+});
