+++ libs/domain/cart/totals/components/subtotal/subtotal.schema.ts
@@ -1,0 +1,6 @@
+import { ContentComponentSchema } from '@spryker-oryx/experience';
+
+export const cartTotalsSubtotalComponentSchema: ContentComponentSchema = {
+  name: 'Cart totals subtotal',
+  group: 'Cart',
+};
