+++ libs/domain/cart/totals/components/tax/index.ts
@@ -1,0 +1,2 @@
+export * from './tax.component';
+export * from './tax.schema';
