+++ libs/domain/cart/totals/components/tax/tax.schema.ts
@@ -1,0 +1,6 @@
+import { ContentComponentSchema } from '@spryker-oryx/experience';
+
+export const cartTotalsTaxComponentSchema: ContentComponentSchema = {
+  name: 'Cart totals tax',
+  group: 'Cart',
+};
