+++ libs/domain/cart/totals/components/total/index.ts
@@ -1,0 +1,4 @@
+export * from './total.component';
+export * from './total.model';
+export * from './total.schema';
+export * from './total.styles';
