+++ libs/domain/cart/totals/components/total/total.model.ts
@@ -1,0 +1,3 @@
+export interface CartTotalsTotalOptions {
+  enableTaxMessage?: boolean;
+}
