+++ libs/domain/cart/totals/components/total/total.styles.ts
@@ -3,71 +3,6 @@ import { smScreen } from '@spryker-oryx/themes/breakpoints';
 import { HeadingTag, headingUtil } from '@spryker-oryx/ui/heading';
 import { css } from 'lit';
 
-const cartTotalRules = css`
-  :host {
-    display: flex;
-    justify-content: space-between;
-    padding-inline: var(--oryx-space-4);
-  }
-
-  span {
-    ${headingUtil(HeadingTag.H6)};
-
-    padding-block: var(--oryx-space-2);
-  }
-`;
-
-const smallScreenRules = css`
-  :host {
-    display: flex;
-    justify-content: space-between;
-    padding-inline: var(--oryx-space-4);
-    padding-block: var(--oryx-space-2);
-  }
-
-  span {
-    ${headingUtil(HeadingTag.H3)};
-  }
-`;
-
-export const cartTotalsStyles: ThemeStylesheets = [
-  cartTotalRules,
-  { media: smScreen, css: smallScreenRules },
-];
-
-const cartTotalsDiscountRules = css`
-  oryx-collapsible {
-    padding-block: var(--oryx-space-2);
-    box-sizing: border-box;
-    width: 100%;
-  }
-
-  span {
-    padding-block: 0;
-  }
-
-  ul {
-    list-style: none;
-    padding: 0;
-    margin: var(--oryx-space-2) 0 0 0;
-    color: var(--oryx-color-neutral-300);
-  }
-
-  li {
-    display: flex;
-    justify-content: space-between;
-  }
-
-  span:nth-child(2) {
-    margin-inline-start: auto;
-    color: var(--oryx-color-highlight-400);
-  }
-`;
-
-export const cartTotalsDiscountStyles: ThemeStylesheets = [
-  cartTotalsDiscountRules,
-];
-
 const cartTotalsTotalRules = css`
   :host {
     display: flex;
