+++ libs/domain/cart/totals/src/totals.def.ts
@@ -10,4 +10,4 @@ export const cartTotalsComponent = componentDef({
   ],
 });
 
-export * from '../components/cart-total.def';
+export * from '../components/component.def';
